hooks and content provider added
