+++ src/App.jsx
@@ -1,4 +1,3 @@
-import { useState } from 'react'
 import Header from './components/Header';
 import Banner from './components/Banner';
 import Learnmore from './components/Learnmore';
@@ -9,10 +8,13 @@ import Latestnews from './components/Latestnews';
 import Readmore from './components/Readmore';
 import Logosec from './components/Logosec';
 import Footer from './components/Footer';
+import DataPorvider from './providers/DataPorvider';
 function App() {
+
 
   return (
     <>
+  <DataPorvider>
       <Header/>
       <Banner/>
       <Learnmore/>
@@ -23,6 +25,7 @@ function App() {
       <Readmore/>
       <Logosec/>
       <Footer/>
+      </DataPorvider>
     </>
