+++ src/contexts/DataContext.js
@@ -1,0 +1,4 @@
+import { createContext } from "react";
+
+  const DataContext = createContext();
+  export default DataContext;
