+++ src/hooks/useData.js
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import DataContext from "../contexts/DataContext";
+
+export default function useData(){
+    return useContext(DataContext);
+}
